Got blue line to render

+++ include/drawpp.hpp
@@ -18,8 +18,8 @@ void setTitle(const char* title);
 // DShape
 // Draws given shape
 void shape(DShape _shape);
-void shape(DShape _shape, int x,int  y);
-void shape(DShape _shape, int x, int y, int w, int h);
+void shape(DShape _shape, int x, int  y);
+void shape(DShape _shape, int x, int y, float w, float h);
 DShape loadShape(std::string filename);
 DShape createShape();
 DShape createShape(DShape::ShapeType type);
